helper_scripts/requirements.txt
numpy
matplotlib
argparse
pillow